Add files via upload
Face Recognition System-It is a python project using open CV in which first of all we did face detection and then passed that detected face to a face recognition model to get a confidence score near to the person present in database of input image. When the person crosses the thresold value, the lock gets unlocked and user is recognized.